refactor: move byePhrase to separate module
