+++ src/operations/getHash.js
