+++ src/operations/getList.js
