+++ src/operations/getOs.js
