+++ src/operations/goToDir.js
