+++ src/utils/getByePhrase.js
@@ -1,0 +1,2 @@
+export const getByePhrase = (name) =>
+  `\nThank you for using File Manager, ${name || "Dear Checker"}, goodbye!`;
